Log UUID inclusion, fixes #23 (#26)
Added the ability to include/exclude the hostname and a UUID in the filename. This will solve possible race conditions when running multiple instances of the plugin dumping into the same bucket, as well as allow a glob with a single pattern to match files (useful for uploading to BigQuery).

On a technical side the file name generation code has been broken out and is now thread-safe which takes us part of the way toward fixing #2.